Simplify code, no longer care if networkzone is restricted.
