+++ pkg/controllers/dynakube/oneagent/oneagent_reconciler.go
@@ -111,8 +111,8 @@ func (r *Reconciler) Reconcile(ctx context.Context, dk *dynakube.DynaKube, dtCli
 		}
 	}
 
-	if errors.Is(err, oaconnectioninfo.StaleRestrictedNetworkZoneError) { // This only informational
-		log.Info("OneAgent endpoints in a restricted network-zone do not yet match the local ActiveGate Service IP, postponing OneAgent deployment until the ActiveGate has re-registered", "dynakube", dk.Name)
+	if errors.Is(err, oaconnectioninfo.StaleNetworkZoneEndpointsError) { // This only informational
+		log.Info("OneAgent endpoints do not yet advertise every local ActiveGate Service IP, postponing OneAgent deployment until the ActiveGate has re-registered", "dynakube", dk.Name)
 	}
 
 	if err != nil {
